factor out file size helper
