+++ aurora/bag_transfer/lib/view_helpers.py
@@ -1,0 +1,5 @@
+def file_size(self, num):
+    for unit in ['B', 'KB', 'MB', 'GB']:
+        if abs(num) < 1024.0:
+            return "%3.1f %s" % (num, unit)
+        num /= 1024.0
